rating type changed from int to float

+++ backend/movies/src/main/java/com/javact/movies/dto/ReviewDto.java
@@ -10,11 +10,11 @@ public class ReviewDto {
     private Long id;
     private String email;
     private String username;
-    private int rating;
+    private Float rating;
     private String comment;
     private Date fullDate;
 
-    public ReviewDto(Long id, String email,String username, int rating, String comment, Date fullDate) {
+    public ReviewDto(Long id, String email, String username, Float rating, String comment, Date fullDate) {
         this.id = id;
         this.email = email;
         this.username = username;
